django_testing/requirements.txt
django
djangorestframework
django-filter
psycopg2-binary
model_bakery

